storm: add vpd, cryptohome and xz-utils for test image
We are currently in a transition of using the
install_factory_toolkit.run (generated by the ./make_factory_toolkit.sh
under src/scripts) to turn a test image into a factory image, because of
that, packages needed by factory need to be in the test image.

This CL try to add the packages needed by factory. More specifically:
 - vpd for writing information like mac address.
 - cryptohome to verify the TPM EK until a new conclusion from issue
   31800.
 - xz-utils for archving the reports.

BUG=chrome-os-partner:31684, chrome-os-partner:31800
TEST=cbuildbot --remote -g 215662 storm-release

Change-Id: Ifcab95c7ff12cd66c4dd67e7c1eb240c9fe9462c
Reviewed-on: https://chromium-review.googlesource.com/215662
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Chun-ta Lin <itspeter@chromium.org>
Commit-Queue: Chun-ta Lin <itspeter@chromium.org>
Tested-by: Chun-ta Lin <itspeter@chromium.org>