chromeos-chrome: Revert to just before the chromium-source dep was added.
The chromium-source workflow still has a couple of bugs that need to
be ironed as witnessed by people using chromeos-chrome to build form
source (apologies!).

Instead of disrupting existing chromeos-chrome workflows revert to
what we used to do before. This way we postpone the bug-fixing to when
we add another package using chromium-source (likely the mojo
packages).

This CL only changes the 9999 ebuild - the updating of other ebuilds
(including prebuilts) will be done by the PFQ on the next run.

BUG=chromium:410903
TEST=build_packages with chromeos-chrome cros_workon'ed works

Change-Id: I459f229f0a85ae2566d5e123d96c0052649ac88b
Reviewed-on: https://chromium-review.googlesource.com/216334
Reviewed-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>
Commit-Queue: David Zeuthen <zeuthen@chromium.org>
Tested-by: David Zeuthen <zeuthen@chromium.org>