xorg-conf: Remove xorg server dep with USE=-X
We still want the config files (for now, as a way of sharing touchpad
device tuning parameters) but not the server.

This fixes pulling xorg drivers into the freon build, some of which
don't compile properly with our build flags (radeon). The generic
build pulls in a number of extra video drivers.

BUG=chromium:403091
TEST=build_packages --board amd64-generic

Change-Id: I286ecf26afd4f8054fd48053b393c951f5f81c7d
Reviewed-on: https://chromium-review.googlesource.com/217090
Reviewed-by: Daniel Nicoara <dnicoara@chromium.org>
Reviewed-by: Peter Mayo <petermayo@chromium.org>
Tested-by: Michael Spang <spang@chromium.org>
Commit-Queue: Michael Spang <spang@chromium.org>