Update Noto-CJK fonts
A number of known issues in 1.000 have been fixed.
The font family name and postscript name (the latter does not
affect CrOS much except for the PDF generation) have been simplified.

In addition, local.conf in fontconfig is updated to take care of
the family name change.

Also, put 'Noto Sans CJK SC' at the top of the fallback font
list when lang is NOT specified. It's the behavior we had before
Noto Sans CJK was added so that this CL goes back to the previous
behavior. See http://crbug.com/408199 as to why we have to revert to
that.

Two CLs will be accompanied by this change.

 Chromium: https://codereview.chromium.org/551963003/
 Skia: https://codereview.chromium.org/554943002/

BUG=chromium:412151,chromium:408199
TEST='fc-list "Noto Sans CJK"' will have 'Noto Sans CJK {JP,KR,SC,TC}.
TEST=`fc-match :lang={ja,zh-cn,zh-tw,ko}` will return 'Noto Sans CJK
{JP,SC,TC,KR}.

Change-Id: If2c499e0932a083d1309e2925d4b9b50cb37b572
Reviewed-on: https://chromium-review.googlesource.com/217290
Reviewed-by: Daniel Erat <derat@chromium.org>
Tested-by: Jungshik Shin <jshin@chromium.org>
Commit-Queue: Jungshik Shin <jshin@chromium.org>