Make sys-boot/syslinux a DEPEND for base images and RDEPEND for dev images.
The syslinux binaries are required to setup the sysboot legacy BIOS
boot mechanism. This is done at build time and is not required to be
installed in the rootfs while booting a Chromium OS image.

If an already booted Chromium OS base image needs to be modified to
boot from a different partition using syslinux (for example, booting
with Ctrl+L to boot an Ubuntu or a factory test image running
something else), then we assume the device is in dev-mode and we can
use dev_install to install syslinux at that point.

This patch moves sys-boot/syslinux from RDEPEND to DEPEND on x86 and
amd64 boards for base images, and installs it in dev images by making
it a RDEPEND on virtual/target-chromium-os-dev.

BUG=chromium:417123
TEST=cbuildbot panther-embedded_minimal works. Generated image has /syslinux in partition 12.

Change-Id: Id5cd8eebc29a060a6391482c88d52214d2a3aef8
Reviewed-on: https://chromium-review.googlesource.com/219894
Reviewed-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>