coreboot: handle properly config files without newline in the last line
This issue keeps coming back: to generate a UART console enabled
image, the ebuild concatenates the target's fwserial config with its
main config file. But when the main config file does not have a
newline in the last line, the concatenation extends the last line,
instead of adding a new line. This results in a garbled config, and
the UART port does not get enabled.

Let's just insert a newline before concatenating the fwserial config,
this will make sure the combined config is valid.

BUG=none
TEST=built images for cosmos (which does not have a newline in the
     last line of the config) and storm (which does have the newllne).
     Verified that in both cases the serial interface was properly
     enabled.

Change-Id: Idf486234807a0426b23c9fd6e01aedcb0724be99
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/223574
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>