xorg-server: Skip SetCursor when vtSema is FALSE
Consider below sequence -
1) Cursor is removed : isUp will be FALSE if HW cursor is set.
2) VT switched away from X : vtSema becomes FALSE.
3) xf86CursorSetCursor is called with non-null CursorPtr :
Saves the passed in CursorPtr, fallbacks to SW cursor and invokes
spriteFuncs->SetCursor which saves the area under cursor and restores
the cursor. This sets isUp to TRUE and as vtSema is FALSE saved data
is garbage.
4) VT switched to X : vtSema becomes TRUE. xf86Cursor enable fb access
is called which will remove the SW cursor, i.e copies saved data in #3
to screen.

This results to momentary garbage data on screen. Hence when !vtSema
skip spriteFuncs->SetCursor.

BUG=chrome-os-partner:32488
TEST=Tap touch panel to hide mouse cursor then close lid to enter S3.
When wake up, the mouse cursor should not show garbage.

Change-Id: Iffbc8c8183d5a9765d0fb29a92bf4a84671546fb
Signed-off-by: Yogish Kulkarni <yogishk@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/224645
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>