Turn package.mask into a directory with specialized masking files in it. (1/2)
Portage allows having either a package.mask file, or a package.mask directory
with individual files in it which are smooshed together to create a masking
file. To make automated manipulation of the package masks easier, this change
breaks up the existing mask file into three new files, one for gperf, one for
syslinux, and one for chrome.

BUG=chromium:399356
TEST=Ran "emerge-lumpy -s gperf" before and after the change and verified that
version 3.0.3 was considered the latest available. Temporarily removed the mask
and saw that the version changed to 3.0.4, showing that the split up mask file
was still masking packages.

Change-Id: I4d8230580fd21f70a707257bd4021a2d423eee7f
Reviewed-on: https://chromium-review.googlesource.com/224832
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>