GCC 4.9: Disable call-order reorder file.
With GCC 4.8.2 we were passing a reorder file to the linker to layout
functions in the binary based on call order. This mechanism is
cumbersome and the orderfile gets out-of-date easily. Besides, the
collection process does not work on ChromeOS anymore.

Now that we are using AFDO, we can layout the binary based on the
function hotness information from AFDO.

BUG=chromium:423463
    chromium:406249
TEST=Lots of testing has been done with this in combination with
     GCC 4.9 roll.
     See https://goto.google.com/gcc_49_testing
CQ-DEPEND=CL:217962

Change-Id: Ia1b892c423e57a9e6133d7e8297e48ed0d8534d8
Reviewed-on: https://chromium-review.googlesource.com/225310
Reviewed-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Luis Lozano <llozano@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>