media-libs/waffle: uprev to 1.4.0
In 1.4.0 waffle added 'wflinfo,' a tool similar to glxinfo but able to
run on more platforms as it can use glx or egl.  Using wflinfo will help
with porting chrome os graphics tests to new platforms, and with reducing
platform-specific checks.
There's a small patch on the waffle build here because it uses a cmake
feature not supported in the present version of chrome os cmake.

BUG=chromium:427706
TEST=gmerge waffle to lumpy and run "wflinfo -p glx -a gl"

Change-Id: Iabc48829e3089a24777bb090450216d1db4d2af5
Reviewed-on: https://chromium-review.googlesource.com/225643
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Frank Henigman <fjhenigman@chromium.org>
Tested-by: Frank Henigman <fjhenigman@chromium.org>