libchrome: Backport AlarmTimer from chrome
crrev.com/26436e402e6e1b780f6565539ae0acb3e2976fb9 added a new timer
class to chrome that is capable of waking the system from suspend.
shill and powerd also need to use this timer but unfortunately the git
mirror for the directory where the timer class lives has not been set up
yet and it's probably too close to the branch point now to risk trying
to uprev libchrome anyway.  Instead we can backport that particular CL
into the chrome os tree until the next time we update libchrome.

BUG=chrome-os-partner:31970,chrome-os-partner:32270,chromium:428214
TEST=manual
Built libchrome and checked the build output to verify that the class
was built as a shared library.

Change-Id: I1706ce4eccbc276d2632bbb57c22a0786cd42fa5
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/226183
Reviewed-by: Alex Vakulenko <avakulenko@chromium.org>