media-libs/mesa: fix gbm_create_device
Because dri driver libs need libglapi but do not link it in, gbm apps
(such as chrome with ozone gbm backend, or a waffle app with gbm backend)
have do a hacky dlopen("libglapi.so") or else fail in gbm_create_device()
when it dlopens the driver lib.
This patch moves the hack inside mesa which is much better place for it
as libglapi is a mesa implementation detail users should not rely on.
The long term solution would be to fix the driver linkage, but upsteam
advice is to use this temporary fix for now.
https://freedesktop.org/patch/35642/

BUG=chromium:428985
TEST=emerge-link_freon mesa && emerge-link_freon waffle and run "wflinfo -p gbm -a gles2" on device
TEST=emerge-lumpy mesa && emerge-lumpy waffle and run "wflinfo -p glx -a gl" on device

Change-Id: I701d8a77d5153312527b28eccc89af0db9996be1
Reviewed-on: https://chromium-review.googlesource.com/226730
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Frank Henigman <fjhenigman@chromium.org>
Tested-by: Frank Henigman <fjhenigman@chromium.org>