libchrome:  Backport accessor function change for AlarmTimer class
The repo for the timers directory already exists at
https://chromium.googlesource.com/chromium/src/components/timers but it
uses c++11 so we can't pull it in directly yet.

Meanwhile, powerd needs this change to function correctly so make the
change to the header in the files/ directory of libchrome.

BUG=chromium:428214,chrome-os-partner:31970
TEST=manual
emerge-$BOARD libchrome and check
/build/$BOARD/usr/include/base-293518/components/timers/alarm_timer.h to
verify that the change made it into the header.

Change-Id: I857e9b6570e36867841bdec9133f0a0a45d189e6
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/227460