libchrome: Pull in MockTimer for testing
CL #226668 wants to use MockTimer from Chrome to simulate a timer
in the asynchronous dark resume setup tests, but MockTimer is
currently not compiled into the libchrome that we use. Instead of
duplicating or reimplementing the class, let's just pull it in.

BUG=chrome-os-partner:31970
TEST=manual
Built libchrome and verified that a new library was generated.

Change-Id: I70b085bc2206d39ff5a49e59d3115b1e77f3359d
Signed-off-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/227686
Reviewed-by: Mike Frysinger <vapier@chromium.org>