chromeos-ec: Add ryu_p1 as a board target
We are making the transition from the old EC (ryu_p1) to the new EC
(ryu) for Ryu. During the transition, to prevent breaking workflow for
developers, we need a way for them to build AP/EC firmware for ryu_p1.
Now that ryu_p1 is a board target, one can do:
  EC_FIRMWARE=ryu_p1 emerge-rush_ryu chromeos-ec ... chromeos-bootimage
to build an AP image with ryu_p1 EC firmware packaged inside.

CQ-DEPEND=CL:229340
BUG=chrome-os-partner:33583
TEST=emerge chromeos-ec with and without EC_FIRMWARE=ryu_p1. Check
resulting image.

Change-Id: If2168cd9e28184bdb8239d9cf946c949dbe23f85
Signed-off-by: Vic Yang <victoryang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/229341
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>