Enable lzo and xattr in sys-fs/squashfs-tools
sys-fs/squashfs-tools is installed in the host system to support
creating and decompressing a squashfs filesystem (but not mount it).
These tools require lzo support and xattr (if the rootfs has files
with xattr) in the host and they are not installed in the device
image since the kernel mounts the filesystem.

This patch also removes the "lzo" flag from the chromeos target that
was not installed and had the wrong package name (squashfs instead of
squashfs-tools).

BUG=chromium:382951
TEST=`sudo emerge sys-fs/squashfs-tools; equery u sys-fs/squashfs-tools` shows the right flags.

Change-Id: I0585cb74fb1b0b6785a128d127bb7f47e40de3a9
Reviewed-on: https://chromium-review.googlesource.com/231550
Commit-Queue: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Gaurav Shah <gauravsh@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>