coreboot: install cbfstools on the host unconditionally
MIPS builder is failing because the cbfstools utility is not available
on the host. It should be installed unconditionally, it is needed to
be present on the host for all architectures to succeed.

This also simplifies the ebuild such that there is no platform
specific definitions which pull in just a single common ebuild. Should
the need for platform specific dependencies arise, this could be added
back.

BUG=chrome-os-partner:31438
TEST=none yet, expecting the builder to succeed building urara images

Change-Id: Iee350c635ea0fbb02ec2e419a1cc3c3b5f552f9d
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/232759
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>