chromeos-activate-date: Require non-empty timestamp file.
tlsdate opens /var/cache/tlsdated/timestamp at startup time but only
writes to it after a time sync has happened.

For activate_date.conf, this leaves open the (unlikely) case in which
a transition from non-existent timestamp file to existent-but-empty
timestamp file is interpreted as a time sync and thus leads to writing
of an incorrect ActivateDate.  This CL prevents that case by requiring
that the timestamp file is not zero-sized for ActivateDate to be
written.

BUG=chromium:439024
TEST=manually tested that empty timestamp doesn't trigger writing of
  ActivateDate anymore
CQ-DEPEND=CL:233380

Change-Id: I5b947b5c2ecdfc5ab801d594d355d2d53d3143af
Reviewed-on: https://chromium-review.googlesource.com/232921
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Thiemo Nagel <tnagel@chromium.org>
Commit-Queue: Thiemo Nagel <tnagel@chromium.org>