chromeos-ec: Add support for asymetric EC image
Ryu Sensor is built with 2 images: ryu_sh_loader and
ryu_sh. The later has only a RW image, which can be up to ~80KB.

Add ryu_sh_loader as a image candidate, use .config in board
to prevent adding ec.RO.bin image if not built.

BUG=chrome-os-partner:33908
TEST=emerge-rush_ryu chromeos-ec produce the right images.
Check no error are reported and that for boards where ec.RO.bin is
produced, it is in build/.../firmware directory.

Change-Id: Id23691d074913932abf7fdc9fb9ea44ddef68f42
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/233233