Add patch to fix interaction between zsh and pid namespaces
zsh does not take control of the process group when it is launched,
causing things like SIGINT to be sent to cros_sdk instead, which kills
the entire shell.  Add a patch to fix this for zsh.  Half of this patch
has already been accepted upstream and the other half is there to
prevent a warning from showing on exit.

BUG=chromium:412484
TEST=manual
% emerge zsh
% cros_sdk -- zsh -l
% ^C
% exit

Verified that zsh does not exit on ^C and that no warning is shown on exit.

Change-Id: I95b2269d82e78ad801e9c7cf87ef88d41435b661
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/236475
Reviewed-by: Mike Frysinger <vapier@chromium.org>