Prepare to upgrade harfbuzz to 0.9.37
Chrome is about to upgrade its harfbuzz to 0.9.37. We have to be in
sync with Chrome. (see https://codereview.chromium.org/810153004/ ).

Add the ebuild file for harfbuzz-0.9.37 from the upstream and set
KEYWORDS to '~*' to prevent 0.9.37 from being built.

Once this is landed, a couple of changes will be made with a revision
bump to r1 before unmasking it.

BUG=chromium:445658
TEST=None (ebuild is masked)

Change-Id: I9bbcf2686e45775d39a2128a1cd6cde1c67064bc
Reviewed-on: https://chromium-review.googlesource.com/238551
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Jungshik Shin <jshin@chromium.org>
Tested-by: Jungshik Shin <jshin@chromium.org>