Enable SATA min_power on AC mode
HSIOPC/GPIO71 is used to control power to VCCHSIO, VCCUSB3PLL
and VCCSATA3PLL in S0. PCH will drive HSIOPC low when all the
high speed I/O controllers (xHCI, SATA, GbE and PCIe) are
in idle.

But there is an additional step needed for SATA to enter low
power mode that its PLL can be shutdown: SATA "min_power"
policy needs to be set.

This is already done by laptop mode tools when the AC->battery
transition happens. We want min_power on AC as well. Thanks to
Sameer Nanda for the pointer.

BUG=chrome-os-partner:35147
TEST=confirmed GPIO71 was toggling after boot up on Paine

Change-Id: Ie5fbd01eee6427887f459b4880be836ba61a988a
Signed-off-by: Wenkai Du <wenkai.du@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/239486
Reviewed-by: Sameer Nanda <snanda@chromium.org>