evtest: import & customize for webplot project
Webplot is a tool uses the touch events captured by evtest to draw
finger traces on Chrome. It is desirable to stop passing gesture
events to Chrome.

The patch grabs the device when --grab command line option is
present. In this way, no other processes are able to access the
device.

A bug was filed upstream as
https://bugzilla.freedesktop.org/show_bug.cgi?id=88572

BUG=chromium:443539
TEST=Check that the cursor is frozen when --grab option is specified.
$ evtest --grab

Change-Id: I688cce77f4dccc50e1e34f518cc3d0d79b45ca43
Reviewed-on: https://chromium-review.googlesource.com/240640
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>