permission_broker: Change unit test to rule_engine_unittest
With the refactor of the PermissionBroker class to be a DBusServiceDaemon,
the rule processing has been extracted to a RuleEngine class.  The unit
testing is isolated to that new class, so rename the unit test file.

BUG=chromium:431841
TEST=`FEATURES=test emerge-panther permission_broker`
CQ-DEPEND=I355756c55e7acf1c59b87025c047171ff8ef8202

Change-Id: I4b15f3f13243a08fd1bf914887d9b8d064023ba0
Reviewed-on: https://chromium-review.googlesource.com/241050
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Stephen Fung <stevefung@chromium.org>
Tested-by: Stephen Fung <stevefung@chromium.org>