Increase the padding in the fit image to fit more dtb files
We were getting an error in arm-generic because it had more than 32
dtb files in it.  Let's raise the limit.  64 ought to be enough for
anyone (and I assume doubling the padding doubles the amount).

This is the same thing we did in CL:61042, but back in the day
everyone thought that 32 would be enough for anyone.  Boy were we
shortsighted back then.

BUG=chrome-os-partner:35238
TEST=Build and boot with >32 dtb files.
TEST=Check that vmlinuz is exactly 1024 bytes bigger.

Change-Id: Ia18f0438381e5dccbb20211e807bdf0e1256b3c8
Reviewed-on: https://chromium-review.googlesource.com/242420
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>