webplot: add webplot-9999.ebuild
Webplot is a finger-drawing tool on browsers for touch devices.
This patch adds the 9999 ebuild file for the package.

Note that webplot imports a "remote" package from touch_firmware_test
with some symlinks. In order to emerge the symlinks, we have to copy
the real files in src_unpack.

In addition, this ebuild also installs a script in a binary path so
that webplot could be launched easily.

Also note that this patch now have a CQ-DEPEND on another patch:
https://chromium-review.googlesource.com/#/c/245674/

BUG=chromium:443539
TEST=emerge and cros deploy webplot. It is emerged and installed
to a machine correctly.
CQ-DEPEND=I95e25e719c6b68f1b22800fd10cab0ea06332487

Change-Id: If18fef228db5be23dc769fcbb4e4c01f616262ca
Reviewed-on: https://chromium-review.googlesource.com/246019
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>