Enable USE=v4l2_codec for all CrOSs platforms.
This flag should be enabled for all CrOS, as we are capable of autodetecting
the presence of V4L2 codec in runtime. The flag cannot be removed though,
because non-ChromeOS builds (linux_chromium_chromeos for example) may fail due
to very old headers on the builders.

TEST=None
BUG=None

Change-Id: Ifa0c4a5509c16b91109650133b040f0ff8f83abc
Reviewed-on: https://chromium-review.googlesource.com/246756
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Pawel Osciak <posciak@chromium.org>
Tested-by: Pawel Osciak <posciak@chromium.org>