Make USE=v4l2_codec default on all CrOS platforms for Chrome in the ebuild.
Make the flag default in the chromeos-chrome ebuild. We are capable
of autodetecting the presence of V4L2 codec in runtime. The flag cannot
be removed though, because non-ChromeOS builds (linux_chromium_chromeos
for example) may fail due to very old headers on the builders.

TEST=None
BUG=None

Change-Id: I894f104442dcfc5fe162ce963d171873afdb5156
Reviewed-on: https://chromium-review.googlesource.com/246757
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Pawel Osciak <posciak@chromium.org>
Commit-Queue: Pawel Osciak <posciak@chromium.org>