dev-install: die if we can't detect the board
dev-install needs to call emerge-${BOARD} to list installed packages.
When BOARD is not defined (board missing in cros-board for example), the
call to emerge fails with an obscur message.
Instead, we should die if the board name is empty with a useful message.

BUG=None
TEST=manual: install dev-install on a board not listed in cros-board.
  dev-install fails cleanly.

Change-Id: I8ba5a80786023e85819f6fa5ed77804f6fc922fc
Reviewed-on: https://chromium-review.googlesource.com/256372
Trybot-Ready: Bertrand Simonnet <bsimonnet@chromium.org>
Tested-by: Bertrand Simonnet <bsimonnet@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Bertrand Simonnet <bsimonnet@chromium.org>