cros-workon.eclass: Further changes to support CROS_WORKON_SRCPATH.
A previous CL introduced the said variable, but overlooked its
significance as an override to CROS_WORKON_PROJECT in
array_vars_autocomplete(). This fixes the shortcoming:

- At least one of the variables must contain some (non-null) value.

- If both are defined, then they should have the same number of values.

- project_count is determined by whichever of the two being defined.

Note that these checks are somewhat relaxed in cases where the array
lengths are 1. This is due to the fact that the default initialization
(assignment of an empty value) results in such length.

BUG=brillo:407
TEST=Correctly identifies violations to the above constraints.

Change-Id: I554edf97756a7c4ec2cc5f050dadc208bb676e8b
Reviewed-on: https://chromium-review.googlesource.com/256380
Tested-by: Gilad Arnold <garnold@chromium.org>
Reviewed-by: Bertrand Simonnet <bsimonnet@chromium.org>
Commit-Queue: Gilad Arnold <garnold@chromium.org>