chromeos-initramfs:  allow for using a custom locale list.
This changes the `emake` options for initramfs building to pass the
setting of $RECOVERY_LOCALES to the emake command line.  This allows
customizing the locales to build into recovery images on a board by
board basis.

The immediate reason for the change is to allow certain boards to
reduce initramfs space by dropping locales.  Long-term, the change
seems likely to be generally useful.

BUG=chromium:464171
TEST=emerge with a customized make.conf setting for $RECOVERY_LOCALES

Change-Id: I6910e337281858aef8e01921233e27b8ec3aeb70
Reviewed-on: https://chromium-review.googlesource.com/256739
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>