cros-workon: Inherit git-2 iff CROS_WORKON_PROJECT is defined.
We inherit git-2.eclass so we can use git-2_src_unpack in
cros-workon.elcass. However, git-2 also pulls in dev-vcs/git (and all of
its dependencies) via DEPEND. These dependencies cannot be removed
because they are needed by other ebuilds that use git-2. Using HDEPEND
would make this more correct, albeit still pull in git (into the chroot)
even if a package doesn't strictly need it.

Fortunately, it looks like we only need git-2 (and git) functionality if
CROS_WORKON_PROJECT is defined. This change ensures that non-PROJECT
packages are not git-table, and does not inherit git-2 in this case.
This allows packages that are not repo projects to emerge without
pulling in git.

This also relocates a call to set_vcsid() and ensures it is only run for
PROJECT packages.

BUG=brillo:407
TEST=cros build of brick package doesn't pull in git
TEST=Trybot w/ full builder

Change-Id: Ic26249495e5ed966bea83d778a5604aeceb7b25b
Reviewed-on: https://chromium-review.googlesource.com/259101
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Gilad Arnold <garnold@chromium.org>
Tested-by: Gilad Arnold <garnold@chromium.org>