opengles-headers: add x11-libs/libX11 dependency
eglplatform.h includes X11/Xlib.h and X11/Xutil.h but it has no dependency
so building waffle for instance may fail if libX11 has not been built
before. build_packages will retry it later, so it works most of the
time but the proper solution is to add this dependency.

BUG=chromium:469297
TEST=emerge-daisy_freon waffle

Change-Id: I91f7f6ac45b1fa287adf114e0580b40ad299ee16
Signed-off-by: Dominik Behr <dbehr@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/261328
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>