Update refcode handling to support RW FSP images
Updated code to handle the RW versions of the FSP image if they are
available from the build.  These will be correctly added to the refcode
region in the RW A/B blocks if required.

Testing this requires that the FSP relocation support be enabled in the
coreboot build to generate the required files.

BRANCH=None
BUG=None
TEST=Strago build and boot

Change-Id: I9ee3b6d5b5942e8b95b08cf51e3f173e2002ea7e
Signed-off-by: Erik Bjorge <erik.c.bjorge@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/263826
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Commit-Queue: Leroy P Leahy <leroy.p.leahy@intel.com>
Tested-by: Leroy P Leahy <leroy.p.leahy@intel.com>