concerto: build AP firmware using a base configuration
Various concerto based boards can share the very same firmware, built
using the urara firmware configuration.

A suggested way of accomplishing this is to create configuration
symlinks for the new boards, pointing at urara configuration files.

There still is the need to actually tell what particular board this
firmware is supposed to run on. It affects some run time choices for
the firmware, including what device tree from the FIT kernel header is
used.

For boards using symlinks for coreboot configuration file the actual
overlay board name is injected into the CBFS to be used as the board
id to provide a way to determine the board at run time.

BRANCH=none
BUG=chrome-os-partner:37593

TEST=modified depthcharge configuration using the same approach;
     modified concerto-private make.conf file (parent for all concerto
     boards) to set firmware to coreboot/depthcharge; ran
     emerge-kennet chromeos-bootimage; burnt the resulting
     image.serial.bin into a Concerto module installed on mBuB; booted
     to kernel login prompt from USB; verified that kennet device tree
     was used to boot:

  bash-4.2# echo $(cat /proc/device-tree/compatible)
  google,kennetimg,pistachio
  bash-4.2#

Change-Id: I48e9dc26b78f7c9f41b6904f72676a21350d3957
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Signed-off-by: Stefan Reinauer <reinauer@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/262888
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>