wpa_supplicant: UPSTREAM: clear blacklist on connect
It has been noticed that the band steering/load balancing of some
multi-AP networks will lead to an ever-growing list of blacklisted
BSSIDs.  This eventually leads to a connection drop when the connection
is pushed to a distant AP.

Fix this issue by clearing the blacklist upon successful connect.

BUG=chromium:473741
TEST=wifi_matfunc

Change-Id: I880654533564546c77b48ec6f23eec97d3e126ef
Reviewed-on: https://chromium-review.googlesource.com/263974
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>