Merge libchrome_crypto into libchrome umbrella library
In order to facilitate easy uprev of libchrome, we need to make sure
that the chrome libraries are properly slotted. libchrome itself is
Ok, but libchrome_crypto is not slotted and thus it is impossible to
make the old and new version co-exist temporarily, which is what needs
to happen in order pull a new version of chrome libraries, since
SDK depends on libchrome now (via chromeos-dbus-bindings) and cannot
be updated immediately to the new version until the uprev stage is
done during the build. The uprev stage, however, requires SDK to be
already built. So during the initial transition, we must have both
old and new version of libchrome installed.

To make it easy, just remove libchrome_crypto and pull its sources
into libchrome meta-library. The crypto portion will be built as
base-crypto-${BASE_VER}.so, so it is still a separate binary.

BUG=None
TEST=./build_packages
CQ-DEPEND=CL:264054,CL:*211848,CL:*211849

Change-Id: Id56a02e6bd8a3828d3b7e8ec6ea8272da091931a
Reviewed-on: https://chromium-review.googlesource.com/264053
Trybot-Ready: Alex Vakulenko <avakulenko@chromium.org>
Tested-by: Alex Vakulenko <avakulenko@chromium.org>
Reviewed-by: Bertrand Simonnet <bsimonnet@chromium.org>
Commit-Queue: Alex Vakulenko <avakulenko@chromium.org>