Add ebuild for updated libchrome ebuilds (r323904)
Due to a problem with CQ, I cannot simply update the ebuilds to pull
a new version of libchrome because it is used for building SDK tools
which are required to do ebuild uprev.

To work around this, I'm first landing the new ebuilds to pull in the
latest revisions of libchrome and once this lands, I'll switch over
to using this revision for all the targets, and then I'll remove the
old version of the ebuilds.

BUG=chromium:472654
TEST=Verified with the following trybots:
     x86-generic-full amd64-generic-full arm-generic-full
     amd64-generic-asan daisy-full nyan-full pre-cq-group
     daisy-release-group sandybridge-release-group
     pineview-release-group

Change-Id: Ic30e3b95ad5324e10f7089e2ae1328404c3145b8
Reviewed-on: https://chromium-review.googlesource.com/263748
Trybot-Ready: Alex Vakulenko <avakulenko@chromium.org>
Tested-by: Alex Vakulenko <avakulenko@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Alex Vakulenko <avakulenko@chromium.org>