coreboot: run make clean before configuring
Emerge moves the entire source directory into the build area and then
starts configuring and building it. In case someone invoked make
locally in the source tree, the local build directory is copied into
the build area and might throw off the make.

To be on the safe side - invoke 'make clean' before trying to
configure the build.

BRANCH=none
BUG=chrome-os-partner:8313
TEST='emerge-storm coreboot' still succeeds.

Change-Id: Ib95cb946d3e935a6172023298b3fb038b63e7f63
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/265044
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>