Enable building with frame pointers for all boards.
The previous change (259838) only changed the default value
of the -fomit-frame-pointer flag which is
overwritten when building with O2 for the x86_64 compiler. This change
will enable frame pointers regardless of O* setting.

Note that the value of this flag for the
x86_32 compiler is not affected by adding -O2.

See bug for more info.

This allows:
ChromeOS-Wide-Profiling to report rich data
Better debuggability of release images
Better crash reports

BUG=chromium:464650
TEST=performance testing for falco, daisy, lumpy, squawks, x86-alex
(previously tested with fno-omit-frame-pointer).
cbuildbot --hwtest of lumpy, x86-alex, falco,
squawks, daisy with this patch which all passed. Ran 'perf -g' on the
squawks trybot image.

Change-Id: I0392b7383309c4107360dccb65dc44a8788b3f72
Reviewed-on: https://chromium-review.googlesource.com/265058
Reviewed-by: Luis Lozano <llozano@chromium.org>
Commit-Queue: Luis Lozano <llozano@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>