sdk: Add dev-lang/go
Required to build some of the tools in our private board.

The nature of the SDK makes it impossible to keep this change to
ourselves.  The SDK is created before we indicate which board we would
like to build, so it's not possible to override it for our board only.

BUG=brillo:823
TEST=cros_sdk --replace; cros_sdk go

Change-Id: I83b62b3aef8f946e6919d596c31a8621dbdd3196
Reviewed-on: https://chromium-review.googlesource.com/265564
Reviewed-by: Rahul Chaudhry <rahulchaudhry@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Richard Larocque <rlarocque@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Richard Larocque <rlarocque@chromium.org>