Backport patch for sending DBus signals synchronously
Previously, libchrome would only send dbus signals asynchronously
because it assumed that the TaskRunner responsible for sending DBus
messages is separate from the application's TaskRunner.  This is not the
case for powerd and asynchronous signals cause an inversion in the
message order when powerd calls ObjectProxy::CallMethodAndBlock after
sending a signal.

This patch has already landed in chrome but since this issue can cause a
user-noticeable delay on resume, we need to backport it until the next
time libchrome is uprevved.

BUG=chromium:472361
TEST=manual
Compile and deploy libchrome and powerd.  Perform a suspend / resume and
see that nothing is broken.

Change-Id: Ia024e48c8b0538d9aeeb74d48bf8c809e32395fa
Reviewed-on: https://chromium-review.googlesource.com/265780
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Commit-Queue: Chirantan Ekbote <chirantan@chromium.org>
Trybot-Ready: Chirantan Ekbote <chirantan@chromium.org>
Tested-by: Chirantan Ekbote <chirantan@chromium.org>