Update DPTF ebuild files
In preparation for upcoming new DPTF releases based on 3.18 kernel,
we now move DPTF ebuild files to virtuals in order to locate the
correct ebuild file based on kernel USE flags. For original 3.10
kernel based overlays, we will continue to use the older DPTF
releases that rely on out-of-tree kernel modules. We also rename
the old DPTF package to "dptf-3_10" as a result. The package name
"dptf" will be reserved for the newer releases that are based on
the upstream kernel thermal drivers.

BUG=chrome-os-partner:40380
TEST=No regression on user experiences
CQ-DEPEND=CL:271679

Change-Id: Ia36949b4f3cbd3065896a6f65e8cf69680c34e3b
Signed-off-by: Brian Bian <brian.bian@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/266534
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>