cros-workon.eclass: Always fetch all git branches.
When cros-workon falls back to git-2 eclass to fetch source, it was
using a hack branch name of " " to try and force fetching of all
branches. This hack didn't work, except for the master branch.

Instead of trying to decide which branch to pull (which is tricky in
many cases, release branches, tryjobs, etc), this CL slightly improves
the pull all branches hack.

The EGIT_ALL_BRANCH config variable is added to the git-2.eclass,
which overrides the fetch command with one that will fetch all remote
branches.

cros-workon.eclass always sets this variable.

BUG=brillo:344
TEST=Local builds:
   ./build_packages --nousepkg --board panther_embedded \
      virtual/target-os virtual/target-os-dev virtual/target-os-test \
      chromeos-base/autotest-all
   Run with outdated .cache (tarred .cache pulled from a builder).
   Run with cleared .cache.

Change-Id: I8a40bbb69c6364113048b8703ba9384ca95791a1
Reviewed-on: https://chromium-review.googlesource.com/266920
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Don Garrett <dgarrett@chromium.org>