gcc: give the right linker directory to clang.
We use realpath to get the linker directory and then give it to clang.
This will give us the directory that contains the actual binary of binutils.
In that directory, there is a 'ld' which is a symbolic link to ld.bfd.
When we use gold as default linker, clang will still use bfd linker. Because
clang uses 'ld' inside that directory. This CL tries to get the directory
that contains 'ld', no matter whether this 'ld' is symbolic or not, as long as
the gcc is using it, clang can use it too.

BUG=chromium:481209
TEST=USE="clang asan" FEATURES="test" emerge-amd64-generic buffet

Change-Id: I1d9df3e7dfe6f848ef521833f592520d700c2bc4
Reviewed-on: https://chromium-review.googlesource.com/267836
Reviewed-by: Han Shen <shenhan@chromium.org>
Commit-Queue: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>